Created the Hour insert and tweaked the tax insert
Still need to create the database properly and link the functions to it, HOERVER it's progressing smoothly.

The next thing I want to focus on is to make the hours fetch from the database dynamic so when hours are deleted the hours array is updated (I could do that with a function that reloads the array when hours are deleted ).

After that I want to focus on getting the Calendar page up and running, or at least to show a basic calendar with information.